added messages about incorrect input
