--- Contacts/Contacts/App.xaml
@@ -1,10 +1,0 @@
-﻿<Application x:Class="Contacts.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:Contacts"
-             StartupUri="MainWindow.xaml"
-             >
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
