--- Contacts/Contacts/App.xaml.cs
@@ -1,13 +1,0 @@
-﻿using System.Configuration;
-using System.Data;
-using System.Windows;
-
-namespace Contacts;
-
-/// <summary>
-/// Interaction logic for App.xaml
-/// </summary>
-public partial class App : Application
-{
-}
-
