--- Contacts/Contacts/MainWindow.xaml
@@ -1,88 +1,0 @@
-﻿<Window x:Class="Contacts.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Contacts"
-        mc:Ignorable="d"
-        Title="Contacts" Height="516" Width="710" MinWidth="730"
-        xmlns:vm="clr-namespace:Contacts.ViewModel">
-
-    <Window.Resources>
-        <vm:BooleanToVisibilityConverter x:Key="BooleanToVisibility"/>
-        <vm:BooleanInverseConverter x:Key="BooleanInverse"/>
-    </Window.Resources>
-    
-    <Grid>
-        <Grid.Resources>
-            <Style TargetType="{x:Type Label}">
-                <Setter Property="Margin" Value="3"/>
-            </Style>
-            <Style TargetType="{x:Type TextBox}">
-                <Setter Property="Margin" Value="3"/>
-            </Style>
-            <Style TargetType="{x:Type Button}">
-                <Setter Property="Margin" Value="3"/>
-                <Setter Property="Height" Value="25"/>
-                <Setter Property="Width" Value="70"/>
-            </Style>
-        </Grid.Resources>
-
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="260*"/>
-            <ColumnDefinition Width="auto"/>
-            <ColumnDefinition Width="540*"/>
-        </Grid.ColumnDefinitions>
-
-        <DockPanel Grid.Column="0">
-            <Grid DockPanel.Dock="Top" Height="30">
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="auto"/>
-                    <ColumnDefinition/>
-                </Grid.ColumnDefinitions>
-
-                <Label Content="Find:" Grid.Column="0"/>
-                <TextBox x:Name="SearchTextBox" Height="20" Grid.Column="1"/>
-            </Grid>
-
-            <Grid DockPanel.Dock="Bottom">
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition/>
-                    <ColumnDefinition/>
-                    <ColumnDefinition/>
-                </Grid.ColumnDefinitions>
-
-                <Button x:Name="AddButton" Content="Add" Grid.Column="0" Command="{Binding AddCommand}"/>
-
-                <Button x:Name="EditButton" Content="Edit" Grid.Column="1" Command="{Binding EditCommand}" CommandParameter="{Binding CurrentContact}"/>
-
-                <Button x:Name="RemoveButton" Content="Remove" Grid.Column="2" Command="{Binding RemoveCommand}" CommandParameter="{Binding CurrentContact}"/>
-            </Grid>
-
-            <ListBox x:Name="ContactsListBox" DisplayMemberPath="Name" ItemsSource="{Binding Contacts, UpdateSourceTrigger=Explicit}" SelectedItem="{Binding CurrentContact}"/>
-        </DockPanel>
-
-        <GridSplitter Grid.Column="1" Width="10" Visibility="Hidden"/>
-
-        <Grid Grid.Column="2" VerticalAlignment="Top">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="25*"/>
-                <ColumnDefinition Width="75*"/>
-            </Grid.ColumnDefinitions>
-
-            <Canvas Width="100" Height="100" 
-                    Background="Aquamarine" Grid.Column="0"/>
-
-            <StackPanel Grid.Column="1">
-                <Label Content="Name:"/>
-                <TextBox x:Name="NameTextBox" Text="{Binding CurrentContact.Name, Mode=TwoWay}" IsReadOnly="{Binding IsEditMode, Converter={StaticResource BooleanInverse}}"/>
-                <Label Content="Phone number:"/>
-                <TextBox x:Name="PhoneTextBox" Text="{Binding CurrentContact.PhoneNumber, Mode=TwoWay}" IsReadOnly="{Binding IsEditMode, Converter={StaticResource BooleanInverse}}"/>
-                <Label Content="Email:"/>
-                <TextBox x:Name="EmailTextBox" Text="{Binding CurrentContact.Email, Mode=TwoWay}" IsReadOnly="{Binding IsEditMode, Converter={StaticResource BooleanInverse}}"/>
-                <Button x:Name="ApplyButton" Content="Apply" HorizontalAlignment="Left" Command="{Binding ApplyCommand}" Visibility="{Binding IsEditMode, Converter={StaticResource BooleanToVisibility}}"/>
-            </StackPanel>
-
-        </Grid>
-    </Grid>
-</Window>
