--- Contacts/Contacts/MainWindow.xaml.cs
@@ -1,16 +1,0 @@
-﻿using Contacts.Model;
-using Contacts.ViewModel;
-using System.Windows;
-namespace Contacts;
-
-/// <summary>
-/// Interaction logic for MainWindow.xaml
-/// </summary>
-public partial class MainWindow : Window
-{
-    public MainWindow()
-    {
-        InitializeComponent();
-        this.DataContext = new MainVM();
-    }
-}
