--- Contacts/View.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.13.35825.156 d17.13
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Contacts", "Contacts\Contacts.csproj", "{49BAAE84-C404-4987-9C73-588B3079EDCE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{49BAAE84-C404-4987-9C73-588B3079EDCE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{49BAAE84-C404-4987-9C73-588B3079EDCE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{49BAAE84-C404-4987-9C73-588B3079EDCE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{49BAAE84-C404-4987-9C73-588B3079EDCE}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {58D34C00-FCD4-4BFA-9C16-7B204D735D35}
-	EndGlobalSection
-EndGlobal
