+++ View/ContactCollection/MainWindow.xaml
@@ -18,9 +18,12 @@
         <vm:BooleanInverseConverter x:Key="BooleanInverse"/>
 
         <DataTemplate x:Key="contactTemplate">
-            <StackPanel Margin="3">
-                <TextBlock Text="{Binding Name}"/>
-                <TextBlock Text="{Binding PhoneNumber}"/>
+            <StackPanel Margin="3" Orientation="Horizontal">
+                <Image Width="40" Height="40" Margin="0,0,8,0" VerticalAlignment="Center"/>
+                <StackPanel>
+                    <TextBlock FontSize="12" FontWeight="SemiBold" Text="{Binding Name}"/>
+                    <TextBlock FontSize="10" Text="{Binding PhoneNumber}"/>
+                </StackPanel>
             </StackPanel>
         </DataTemplate>
     </Window.Resources>
@@ -28,9 +31,6 @@
     <Grid>
         <Grid.Resources>
             <Style TargetType="{x:Type Label}">
-                <Setter Property="Margin" Value="3"/>
-            </Style>
-            <Style TargetType="{x:Type TextBox}">
                 <Setter Property="Margin" Value="3"/>
             </Style>
             <Style TargetType="{x:Type Button}">
