+++ View/ContactCollection/Model/Services/ContactSerializer.cs
@@ -8,18 +8,18 @@ namespace ContactCollection.Model.Services
         /// Предоставляет методы для сериализации и десериализации
         /// коллекции контактов.
         /// </summary>
-        public class ContactSerializer
+        public static class ContactSerializer
         {
             /// <summary>
             /// Путь к файлу, в котором хранятся контакты.
             /// </summary>
-            private string _filePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments), "Contacts", "contacts.json");
+            private static string _filePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments), "Contacts", "contacts.json");
 
             /// <summary>
             /// Инициализирует новый экземпляр класса <see cref="ContactSerializer"/>.
             /// Создаёт директорию, если она не существует.
             /// </summary>
-            public ContactSerializer()
+            public static void IsDirectoryExist()
             {
                 string directoryPath = Path.GetDirectoryName(_filePath);
                 if (!Directory.Exists(directoryPath))
@@ -28,22 +28,22 @@ namespace ContactCollection.Model.Services
                 }
             }
 
-            /// <summary>
-            /// Сохраняет коллекцию контактов.
-            /// </summary>
-            /// <param name="contacts">Коллекция контактов для сохранения.</param>
-            public void SaveContacts(ObservableCollection<Contact> contacts)
+        /// <summary>
+        /// Сохраняет коллекцию контактов.
+        /// </summary>
+        /// <param name="contacts">Коллекция контактов для сохранения.</param>
+            public static void SaveContacts(ObservableCollection<Contact> contacts)
             {
                 string json = JsonConvert.SerializeObject(contacts);
                 File.WriteAllText(_filePath, json);
             }
 
-            /// <summary>
-            /// Хагружает коллекцию контактов из файла.
-            /// </summary>
-            /// <returns>Коллекцию контактов, загруженная из файла.
-            /// Возвращает новую пустую коллекцию, если файл не существует.</returns>
-            public ObservableCollection<Contact> LoadContacts()
+        /// <summary>
+        /// Хагружает коллекцию контактов из файла.
+        /// </summary>
+        /// <returns>Коллекцию контактов, загруженная из файла.
+        /// Возвращает новую пустую коллекцию, если файл не существует.</returns>
+            public static ObservableCollection<Contact> LoadContacts()
             {
                 if (File.Exists(_filePath))
                 {
